Add Insert method for tree
Another implication of this change is the definition of an initialized
and uninitialized tree. Since it's not possible to update a nil pointer
receiver (for good reasons), I've decided to instead go with a similar
approach there is for lists and maps. A tree node that is actually nil
would be considered "uninitialized". Trees now have a Size property that
only has a meaningful value for the root note, and the 0 value for all
other nodes (which could be updated in the future to reflect their
size). A non-nill tree is considered empty if the following conditions are
fulfilled:

- The node has no parent;
- Its size is 0

If both conditions are satisfied, Insert would set a new value to the
root node, rather than adding a new node.